chore: Move all .d.ts files to types folder & lint
