+++ packages/buble/package.json
@@ -19,19 +19,19 @@
     "ci:test": "pnpm run test -- --verbose && pnpm run test:ts",
     "lint": "pnpm run lint:js && pnpm run lint:docs && pnpm run lint:package",
     "lint:docs": "prettier --single-quote --write README.md",
-    "lint:js": "eslint --fix --cache src test",
+    "lint:js": "eslint --fix --cache src test types --ext .js,.ts",
     "lint:package": "prettier --write package.json --plugin=prettier-plugin-package",
     "prebuild": "del-cli dist",
     "prepare": "pnpm run build",
     "prepublishOnly": "pnpm run lint && pnpm run test && pnpm run test:ts",
     "pretest": "pnpm run build",
     "test": "ava",
-    "test:ts": "tsc index.d.ts test/types.ts --noEmit"
+    "test:ts": "tsc types/index.d.ts test/types.ts --noEmit"
   },
   "files": [
     "dist",
-    "index.d.ts",
     "src",
+    "types",
     "README.md"
   ],
   "keywords": [
@@ -46,7 +46,7 @@
     "rollup": "^1.20.0"
   },
   "dependencies": {
-    "@rollup/pluginutils": "^3.0.4",
+    "@rollup/pluginutils": "^3.0.8",
     "@types/buble": "^0.19.2",
     "buble": "^0.19.8"
   },
@@ -64,5 +64,6 @@
       "!**/types.ts"
     ]
   },
-  "module": "dist/rollup-plugin-buble.es.js"
+  "module": "dist/rollup-plugin-buble.es.js",
+  "types": "types/index.d.ts"
 }
