+++ packages/buble/types/index.d.ts
@@ -1,16 +1,17 @@
-import { Plugin } from 'rollup';
+import { FilterPattern } from '@rollup/pluginutils';
 import { TransformOptions } from 'buble';
+import { Plugin } from 'rollup';
 
 export interface RollupBubleOptions extends TransformOptions {
   /**
    * A minimatch pattern, or array of patterns, of files that should be
    * processed by this plugin (if omitted, all files are included by default)
    */
-  include?: string | RegExp | ReadonlyArray<string | RegExp> | null;
+  include?: FilterPattern;
   /**
    * Files that should be excluded, if `include` is otherwise too permissive.
    */
-  exclude?: string | RegExp | ReadonlyArray<string | RegExp> | null;
+  exclude?: FilterPattern;
 }
 
 /**
