+++ packages/commonjs/package.json
@@ -22,7 +22,7 @@
     "ci:test": "pnpm run test -- --verbose && pnpm run test:ts",
     "lint": "pnpm run lint:js && pnpm run lint:docs && pnpm run lint:package",
     "lint:docs": "prettier --single-quote --write README.md",
-    "lint:js": "eslint --fix --cache src test",
+    "lint:js": "eslint --fix --cache src test types --ext .js,.ts",
     "lint:package": "prettier --write package.json --plugin=prettier-plugin-package",
     "prebuild": "del-cli dist",
     "prepare": "pnpm run build",
@@ -49,7 +49,7 @@
     "rollup": "^1.20.0"
   },
   "dependencies": {
-    "@rollup/pluginutils": "^3.0.0",
+    "@rollup/pluginutils": "^3.0.8",
     "estree-walker": "^1.0.1",
     "is-reference": "^1.1.2",
     "magic-string": "^0.25.2",
