+++ packages/json/test/types.ts
@@ -1,23 +1,23 @@
-// @ts-check
-import json, { RollupJsonOptions } from '..';
+import { RollupOptions } from 'rollup';
 
-/** @type {import("rollup").RollupOptions} */
-const config = {
-	input: 'main.js',
-	output: {
-		file: 'bundle.js',
-		format: 'iife'
-	},
-	plugins: [
-		json({
-			include: 'node_modules/**',
-			exclude: ['node_modules/foo/**', 'node_modules/bar/**'],
-			preferConst: true,
-			indent: '  ',
-			compact: true,
-			namedExports: true
-		})
-	]
+import json from '..';
+
+const config: RollupOptions = {
+  input: 'main.js',
+  output: {
+    file: 'bundle.js',
+    format: 'iife'
+  },
+  plugins: [
+    json({
+      include: 'node_modules/**',
+      exclude: ['node_modules/foo/**', 'node_modules/bar/**'],
+      preferConst: true,
+      indent: '  ',
+      compact: true,
+      namedExports: true
+    })
+  ]
 };
 
 export default config;
