+++ packages/run/package.json
@@ -18,14 +18,14 @@
     "ci:test": "pnpm run test -- --verbose",
     "lint": "pnpm run lint:js && pnpm run lint:docs && pnpm run lint:package",
     "lint:docs": "prettier --single-quote --write README.md",
-    "lint:js": "eslint --fix --cache lib test",
+    "lint:js": "eslint --fix --cache lib test types --ext .js,.ts",
     "lint:package": "prettier --write package.json --plugin=prettier-plugin-package",
     "prepublishOnly": "pnpm run lint && pnpm run test",
     "test": "ava"
   },
   "files": [
-    "index.d.ts",
-    "lib"
+    "lib",
+    "types"
   ],
   "keywords": [
     "rollup",
